+++ src/App.css
@@ -1,17 +1,17 @@
 /* General Overall */
 @font-face {
   font-family: "StagSans-BoldItalic";
-  src: url("../resources/StagSans-BoldItalic.ttf") format("truetype");
+  src: url("./resources/fonts/StagSans-BoldItalic.ttf") format("truetype");
 }
 
 @font-face {
   font-family: "Lato";
-  src: url("../resources/Lato-Regular.ttf") format("truetype");
+  src: url("./resources/fonts/Lato-Regular.ttf") format("truetype");
 }
 
 @font-face {
   font-family: "Lato-Bold";
-  src: url("../resources/Lato-Bold.ttf") format("truetype");
+  src: url("./resources/fonts/Lato-Bold.ttf") format("truetype");
 }
 
 html,
@@ -106,7 +106,7 @@ body {
 
 /* Register and Login CSS */
 .background-container {
-  background-image: url("../resources/register-image1.jpeg");
+  background-image: url("./resources/images/register-image1.jpeg");
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
@@ -116,7 +116,7 @@ body {
 }
 
 #login-background-container {
-  background-image: url("../resources/login-image1.jpeg");
+  background-image: url("./resources/images/login-image1.jpeg");
 }
 
 .container {
