+++ src/components/CarouselContent.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import "./components.css";
 
 const CarouselContent = () => {
   return (
