+++ src/pages/ClassTimetable.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef, useState } from "react";
-import "./components.css";
 import {
   DayPilotCalendar,
   DayPilotNavigator,
