+++ src/resources/fonts/StagSans-SemiBoldItalic.ttf
